refactor: moved the 2fa section to a new file & feat(2fa): added auto-submitting for the confirmation of the totp code stage
