+++ components/helpers/user-button/security/security-section.tsx
@@ -1,464 +1,18 @@
+import { User } from "better-auth";
 import { PasswordSection } from "./password-section";
+import { TwoFactorSection } from "./two-factor-section";
 
-export const SecuritySection = () => {
+export const SecuritySection = ({
+  user,
+}: {
+  user: User;
+}) => {
   return (
     <div className="flex md:flex-row flex-col md:gap-0 gap-8 py-3 w-full">
       <p className="text-sm font-medium pointer-events-none">Security</p>
       <div className="flex w-full md:items-end flex-col gap-10">
         <PasswordSection />
-
-        <div className="flex md:w-[72%] flex-col gap-10">
-          <div ref={thirdParent}>
-            {!isTwoFactorBoxOpen && !isRemoveTwoFactorBoxOpen ? (
-              <div className="min-w-[350px]">
-                <DropdownMenu>
-                  <div className="flex items-center justify-between">
-                    <p className="text-sm font-medium">Two-step verification</p>
-                    {!is2FAEnabled ? (
-                      <DropdownMenuTrigger asChild>
-                        <Button
-                          variant={"ghost"}
-                          size={"sm"}
-                          className="text-sm"
-                        >
-                          Add two-step verification
-                        </Button>
-                      </DropdownMenuTrigger>
-                    ) : (
-                      <div className="flex items-center gap-2 min-w-[50%]">
-                        <svg
-                          fill="currentColor"
-                          xmlns="http://www.w3.org/2000/svg"
-                          viewBox="0 0 16 16"
-                          className="text-zinc-700 size-5"
-                        >
-                          <path d="M7 9.5a1.5 1.5 0 1 0 0-3 1.5 1.5 0 0 0 0 3Z"></path>
-                          <path
-                            fillRule="evenodd"
-                            clipRule="evenodd"
-                            d="M4 2c-1.105 0-2 .895-2 2v8c0 1.105.895 2 2 2h8c1.105 0 2-.895 2-2V4c0-1.105-.895-2-2-2H4Zm3 9a3.002 3.002 0 0 0 2.906-2.25H12a.75.75 0 0 0 0-1.5H9.906A3.002 3.002 0 0 0 4 8c0 .941.438 1.785 1.117 2.336A2.985 2.985 0 0 0 7 11Z"
-                          ></path>
-                        </svg>
-                        <p className="text-xs text-zinc-600 mr-2">
-                          Authenticator app
-                        </p>
-                        <Badge variant={"outline"}>Default</Badge>
-                      </div>
-                    )}
-                    {is2FAEnabled && (
-                      <DropdownMenu>
-                        <DropdownMenuTrigger asChild>
-                          <Button
-                            variant={"ghost"}
-                            size="icon"
-                            className="group"
-                          >
-                            <Ellipsis className="h-4 w-4 text-zinc-400 group-hover:text-zinc-800 transition" />
-                          </Button>
-                        </DropdownMenuTrigger>
-                        <DropdownMenuContent
-                          align="center"
-                          className="rounded-lg shadow-lg py-1 px-3 min-w-fit"
-                        >
-                          <DropdownMenuItem
-                            className="cursor-pointer p-0"
-                            onClick={() => {
-                              setIsRemoveTwoFactorBoxOpen(true);
-                              setTwoFactorStage(10);
-                            }}
-                          >
-                            <p className="text-sm text-destructive">Remove</p>
-                          </DropdownMenuItem>
-                        </DropdownMenuContent>
-                      </DropdownMenu>
-                    )}
-                    <DropdownMenuContent
-                      align="center"
-                      className="rounded-lg shadow-lg"
-                      onClick={() => {
-                        setIsTwoFactorBoxOpen(true);
-                      }}
-                    >
-                      <DropdownMenuItem className="cursor-pointer">
-                        <svg
-                          fill="currentColor"
-                          xmlns="http://www.w3.org/2000/svg"
-                          viewBox="0 0 16 16"
-                          className="text-zinc-700"
-                        >
-                          <path d="M7 9.5a1.5 1.5 0 1 0 0-3 1.5 1.5 0 0 0 0 3Z"></path>
-                          <path
-                            fillRule="evenodd"
-                            clipRule="evenodd"
-                            d="M4 2c-1.105 0-2 .895-2 2v8c0 1.105.895 2 2 2h8c1.105 0 2-.895 2-2V4c0-1.105-.895-2-2-2H4Zm3 9a3.002 3.002 0 0 0 2.906-2.25H12a.75.75 0 0 0 0-1.5H9.906A3.002 3.002 0 0 0 4 8c0 .941.438 1.785 1.117 2.336A2.985 2.985 0 0 0 7 11Z"
-                          ></path>
-                        </svg>
-                        <p className="text-sm text-zinc-600">
-                          Authenticator app
-                        </p>
-                      </DropdownMenuItem>
-                    </DropdownMenuContent>
-                  </div>
-                </DropdownMenu>
-              </div>
-            ) : (
-              <>
-                <Card className="shadow-md">
-                  <CardHeader>
-                    <CardTitle className="text-sm tracking-tight">
-                      {twoFactorStage === 1 &&
-                        "First enter your current password"}
-                      {twoFactorStage === 10 &&
-                        "First enter your current password"}
-                      {twoFactorStage === 2 && "Add authenticator application"}
-                      {twoFactorStage === 3 && "Add authenticator application"}
-                      {twoFactorStage === 4 && "Add authenticator application"}
-                    </CardTitle>
-                    {twoFactorStage === 2 && (
-                      <CardDescription className="text-xs">
-                        Set up a new sign-in method in your authenticator app
-                        and scan the following QR code to link it to your
-                        account.
-                      </CardDescription>
-                    )}
-                    {twoFactorStage === 10 && (
-                      <CardDescription className="text-xs">
-                        To remove two-factor authentication, enter your current
-                        password.
-                      </CardDescription>
-                    )}
-                  </CardHeader>
-                  <CardContent>
-                    {twoFactorStage === 1 && (
-                      <>
-                        {error && <ErrorCard size="sm" error={error} />}
-                        <Form {...twoFactorForm}>
-                          <form
-                            className="space-y-6"
-                            onSubmit={twoFactorForm.handleSubmit(
-                              onTwoFactorPasswordSubmit
-                            )}
-                          >
-                            <FormField
-                              control={twoFactorForm.control}
-                              name="currentPassword"
-                              render={({ field }) => (
-                                <FormItem>
-                                  <FormLabel className="text-sm">
-                                    Current Password
-                                  </FormLabel>
-                                  <FormControl>
-                                    <div className="relative">
-                                      <Input
-                                        {...field}
-                                        autoCorrect="off"
-                                        autoComplete="off"
-                                        disabled={isLoading}
-                                        type={
-                                          isPasswordVisible
-                                            ? "text"
-                                            : "password"
-                                        }
-                                        className="pe-9"
-                                      />
-                                      <button
-                                        className="absolute inset-y-0 end-0 flex h-full w-9 items-center justify-center rounded-e-lg text-muted-foreground/80 outline-offset-2 transition-colors hover:text-foreground focus:z-10 focus-visible:outline focus-visible:outline-2 focus-visible:outline-ring/70 disabled:pointer-events-none disabled:cursor-not-allowed disabled:opacity-50"
-                                        type="button"
-                                        onClick={toggleVisibility}
-                                        aria-label={
-                                          isPasswordVisible
-                                            ? "Hide password"
-                                            : "Show password"
-                                        }
-                                        aria-pressed={isPasswordVisible}
-                                        aria-controls="password"
-                                      >
-                                        {isPasswordVisible ? (
-                                          <EyeOff
-                                            size={16}
-                                            strokeWidth={2}
-                                            aria-hidden="true"
-                                          />
-                                        ) : (
-                                          <Eye
-                                            size={16}
-                                            strokeWidth={2}
-                                            aria-hidden="true"
-                                          />
-                                        )}
-                                      </button>
-                                    </div>
-                                  </FormControl>
-                                  <FormMessage />
-                                </FormItem>
-                              )}
-                            />
-                            <Button
-                              size={"sm"}
-                              variant={"ghost"}
-                              type="button"
-                              disabled={isLoading}
-                              className="mt-4 mr-2"
-                              onClick={() => {
-                                setIsTwoFactorBoxOpen(false);
-                                setIsRemoveTwoFactorBoxOpen(false);
-                                twoFactorForm.reset();
-                                setError("");
-                              }}
-                            >
-                              Cancel
-                            </Button>
-                            <Button
-                              size={"sm"}
-                              type="submit"
-                              disabled={isLoading}
-                              className="mt-4"
-                            >
-                              {isLoading && (
-                                <Loader className="mr-1 size-2 text-muted-foreground animate-spin" />
-                              )}
-                              Save
-                            </Button>
-                          </form>
-                        </Form>
-                      </>
-                    )}
-                    {twoFactorStage === 2 && (
-                      <>
-                        <div className="flex flex-col items-center justify-center">
-                          <QrCode
-                            value={totpUri || ""}
-                            className="size-[170px]"
-                          />
-                        </div>
-                        <Button
-                          size={"sm"}
-                          variant={"ghost"}
-                          type="button"
-                          disabled={isLoading}
-                          className="mt-10 mr-2"
-                          onClick={() => {
-                            setIsTwoFactorBoxOpen(false);
-                            twoFactorForm.reset();
-                            setError("");
-                          }}
-                        >
-                          Cancel
-                        </Button>
-                        <Button
-                          size={"sm"}
-                          type="button"
-                          disabled={isLoading}
-                          className="mt-10"
-                          onClick={() => {
-                            setTwoFactorStage(3);
-                          }}
-                        >
-                          {isLoading && (
-                            <Loader className="mr-1 size-2 text-muted-foreground animate-spin" />
-                          )}
-                          Continue
-                        </Button>
-                      </>
-                    )}
-                    {twoFactorStage === 3 && (
-                      <>
-                        {error && <ErrorCard size="sm" error={error} />}
-                        <Form {...totpCodeForm}>
-                          <form
-                            className="space-y-6 flex flex-col items-center justify-center"
-                            onSubmit={totpCodeForm.handleSubmit(
-                              () => onTotpCodeSubmit
-                            )}
-                          >
-                            <FormField
-                              control={totpCodeForm.control}
-                              name="otp"
-                              render={({ field }) => (
-                                <FormItem>
-                                  <FormControl>
-                                    <InputOTP
-                                      disabled={isLoading}
-                                      maxLength={6}
-                                      {...field}
-                                    >
-                                      <InputOTPGroup>
-                                        <InputOTPSlot index={0} />
-                                        <InputOTPSlot index={1} />
-                                        <InputOTPSlot index={2} />
-                                        <InputOTPSeparator className="mx-1 text-zinc-600" />
-                                        <InputOTPSlot index={3} />
-                                        <InputOTPSlot index={4} />
-                                        <InputOTPSlot index={5} />
-                                      </InputOTPGroup>
-                                    </InputOTP>
-                                  </FormControl>
-                                  <FormMessage />
-                                </FormItem>
-                              )}
-                            />
-                            <div className="self-start">
-                              <Button
-                                size={"sm"}
-                                variant={"ghost"}
-                                type="button"
-                                disabled={isLoading}
-                                className="mt-4 mr-2"
-                                onClick={() => {
-                                  setIsTwoFactorBoxOpen(false);
-                                  twoFactorForm.reset();
-                                  totpCodeForm.reset();
-                                  setTwoFactorStage(1);
-                                  setError("");
-                                }}
-                              >
-                                Cancel
-                              </Button>
-                              <Button
-                                size={"sm"}
-                                type="button"
-                                disabled={isLoading}
-                                onClick={() => {
-                                  onTotpCodeSubmit({
-                                    otp: totpCodeForm.getValues().otp,
-                                  });
-                                }}
-                                className="mt-4"
-                              >
-                                {isLoading && (
-                                  <Loader className="mr-1 size-2 text-muted-foreground animate-spin" />
-                                )}
-                                Continue
-                              </Button>
-                            </div>
-                          </form>
-                        </Form>
-                      </>
-                    )}
-                    {twoFactorStage === 4 && (
-                      <div className="flex flex-col">
-                        <CardDescription className="text-xs">
-                          Two-step verification is now enabled. When signing in,
-                          you will need to enter a verification code from this
-                          authenticator as an additional step.
-                        </CardDescription>
-                        <Button
-                          size={"sm"}
-                          type="button"
-                          disabled={isLoading}
-                          className="mt-4 self-end"
-                          onClick={() => {
-                            setIsTwoFactorBoxOpen(false);
-                            twoFactorForm.reset();
-                            totpCodeForm.reset();
-                            setError("");
-                          }}
-                        >
-                          Finish
-                        </Button>
-                      </div>
-                    )}
-                    {twoFactorStage === 10 && (
-                      <>
-                        {error && <ErrorCard size="sm" error={error} />}
-                        <Form {...removeTwoFactorForm}>
-                          <form
-                            className="space-y-6"
-                            onSubmit={removeTwoFactorForm.handleSubmit(
-                              onRemoveTwoFactorSubmit
-                            )}
-                          >
-                            <FormField
-                              control={removeTwoFactorForm.control}
-                              name="currentPassword"
-                              render={({ field }) => (
-                                <FormItem>
-                                  <FormLabel className="text-sm">
-                                    Current Password
-                                  </FormLabel>
-                                  <FormControl>
-                                    <div className="relative">
-                                      <Input
-                                        {...field}
-                                        autoCorrect="off"
-                                        autoComplete="off"
-                                        disabled={isLoading}
-                                        type={
-                                          isPasswordVisible
-                                            ? "text"
-                                            : "password"
-                                        }
-                                        className="pe-9"
-                                      />
-                                      <button
-                                        className="absolute inset-y-0 end-0 flex h-full w-9 items-center justify-center rounded-e-lg text-muted-foreground/80 outline-offset-2 transition-colors hover:text-foreground focus:z-10 focus-visible:outline focus-visible:outline-2 focus-visible:outline-ring/70 disabled:pointer-events-none disabled:cursor-not-allowed disabled:opacity-50"
-                                        type="button"
-                                        onClick={toggleVisibility}
-                                        aria-label={
-                                          isPasswordVisible
-                                            ? "Hide password"
-                                            : "Show password"
-                                        }
-                                        aria-pressed={isPasswordVisible}
-                                        aria-controls="password"
-                                      >
-                                        {isPasswordVisible ? (
-                                          <EyeOff
-                                            size={16}
-                                            strokeWidth={2}
-                                            aria-hidden="true"
-                                          />
-                                        ) : (
-                                          <Eye
-                                            size={16}
-                                            strokeWidth={2}
-                                            aria-hidden="true"
-                                          />
-                                        )}
-                                      </button>
-                                    </div>
-                                  </FormControl>
-                                  <FormMessage />
-                                </FormItem>
-                              )}
-                            />
-                            <Button
-                              size={"sm"}
-                              variant={"ghost"}
-                              type="button"
-                              disabled={isLoading}
-                              className="mt-4 mr-2"
-                              onClick={() => {
-                                setIsTwoFactorBoxOpen(false);
-                                setIsRemoveTwoFactorBoxOpen(false);
-                                twoFactorForm.reset();
-                                setError("");
-                              }}
-                            >
-                              Cancel
-                            </Button>
-                            <Button
-                              size={"sm"}
-                              type="submit"
-                              disabled={isLoading}
-                              className="mt-4"
-                            >
-                              {isLoading && (
-                                <Loader className="mr-1 size-2 text-muted-foreground animate-spin" />
-                              )}
-                              Save
-                            </Button>
-                          </form>
-                        </Form>
-                      </>
-                    )}
-                  </CardContent>
-                </Card>
-                <div ref={twoFactorBoxBottomRef} />
-              </>
-            )}
-          </div>
-        </div>
+        <TwoFactorSection user={user} />
 
         <div className="flex md:w-[72%] flex-col gap-10">
           <div
